+++ bin/BuildConsensus.py
@@ -20,7 +20,7 @@ from presto.Defaults import default_delimiter, default_barcode_field, \
 from presto.Commandline import CommonHelpFormatter, checkArgs, getCommonArgParser, parseCommonArgs
 from presto.Annotation import flattenAnnotation, mergeAnnotation, getAnnotationValues, \
                               annotationConsensus
-from presto.IO import getFileType, printLog
+from presto.IO import getFileType, printLog, printWarning
 from presto.Sequence import subsetSeqSet, calculateDiversity, \
                             qualityConsensus, frequencyConsensus, indexSeqSets, \
                             calculateSetError, deleteSeqPositions, findGapPositions
@@ -213,12 +213,12 @@ def processBCQueue(alive, data_queue, result_queue, cons_func, cons_args={},
             # Feed results to result queue
             result_queue.put(result)
         else:
-            sys.stderr.write('PID %s:  Error in sibling process detected. Cleaning up.\n' \
+            sys.stderr.write('PID %s> Error in sibling process detected. Cleaning up.\n' \
                              % os.getpid())
             return None
     except:
         alive.value = False
-        sys.stderr.write('Error processing sequence set with ID: %s\n' % data.id)
+        printError('Processing sequence set with ID: %s' % data.id, exit=False)
         raise
     
     return None